chnage data type and allow null

+++ src/database/models/products_sales.js
@@ -15,7 +15,7 @@ id_products_sales:{
 },
 
     quantity_product: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.TINYINT,
         allowNull: false
     },
 }, {
